yea swagger fixed and moved around...there is also swagger_spec available that lists swagger info
