+++ next/pages/api/swagger_spec.ts
@@ -1,0 +1,7 @@
+import { withSwagger } from 'next-swagger-doc';
+
+import swaggerSpec from '../../swaggerSpec';
+
+const swaggerHandler = withSwagger(swaggerSpec);
+
+export default swaggerHandler();
