+++ next/pages/documentation/api.tsx
@@ -1,0 +1,26 @@
+import { GetStaticProps, InferGetStaticPropsType } from 'next';
+import { createSwaggerSpec } from 'next-swagger-doc';
+import dynamic from 'next/dynamic';
+import 'swagger-ui-react/swagger-ui.css';
+
+import swaggerSpec from '../../swaggerSpec';
+
+const SwaggerUI = dynamic<{
+  spec: any;
+}>(import('swagger-ui-react'), { ssr: false });
+
+function ApiDoc({ spec }: InferGetStaticPropsType<typeof getStaticProps>) {
+  return <SwaggerUI spec={spec} />;
+}
+
+export const getStaticProps: GetStaticProps = async () => {
+  const spec: Record<string, any> = createSwaggerSpec(swaggerSpec);
+
+  return {
+    props: {
+      spec,
+    },
+  };
+};
+
+export default ApiDoc;
